ML-AW-main-1/code/requirements.txt
torch>=1.8.0
torchvision >= 0.9.0
numpy >= 1.20.3
tqdm
tensorboard>=2.7.0
pyro-ppl>=1.7.0
